fe: refactor api url
