+++ be/requirements.txt
@@ -8,11 +8,19 @@ python-dotenv
 pydantic-settings
 pydantic
 pydantic[email]
-langchain
-langchain-google-genai
 ultralytics
 opencv-python
 Pillow
 numpy
 python-multipart
 alembic
+langchain
+langchain-google-genai
+llama-index
+llama-index-vector-stores-postgres
+llama-index-storage-docstore-postgres
+llama-index-storage-index-store-postgres
+llama-index-vector-stores-qdrant
+llama-index-postprocessor-cohere-rerank
+llama-index-embeddings-cohere
+qdrant-client
