+++ fe/lib/fracture-api.ts
@@ -2,24 +2,6 @@
 
 const API_BASE = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000";
 
-export async function uploadFractureImage(file: File, token: string) {
-  const formData = new FormData();
-  formData.append('file', file);
-  
-  const res = await fetch(`${API_BASE}/api/fracture/upload`, {
-    method: "POST",
-    headers: { Authorization: `Bearer ${token}` },
-    body: formData,
-  });
-  
-  if (!res.ok) {
-    const errorData = await res.json().catch(() => ({ detail: "Upload failed" }));
-    throw new Error(errorData.detail || "Upload failed");
-  }
-  
-  return res.json();
-}
-
 export async function submitStudentAnnotations(
   predictionId: number, 
   annotations: Array<{
@@ -29,6 +11,7 @@ export async function submitStudentAnnotations(
     y_max: number;
     width: number;
     height: number;
+    fracture_type?: string;
     notes?: string;
   }>, 
   token: string
