+++ fe/lib/upload.ts
@@ -1,0 +1,21 @@
+// APIs to handle user upload 
+
+const API_BASE = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000";
+
+export async function uploadFractureImage(file: File, token: string) {
+  const formData = new FormData();
+  formData.append('file', file);
+  
+  const res = await fetch(`${API_BASE}/upload/image`, {
+    method: "POST",
+    headers: { Authorization: `Bearer ${token}` },
+    body: formData,
+  });
+  
+  if (!res.ok) {
+    const errorData = await res.json().catch(() => ({ detail: "Upload failed" }));
+    throw new Error(errorData.detail || "Upload failed");
+  }
+  
+  return res.json();
+}
